+++ src/main/java/dinkplugin/DinkPluginConfig.java
@@ -170,10 +170,24 @@ public interface DinkPluginConfig extends Config {
     }
 
     @ConfigItem(
+        keyName = "screenshotScale",
+        name = "Screenshot Scale",
+        description = "Resizes screenshots in each dimension by the specified percentage.<br/>" +
+            "Useful to avoid Discord's max upload size of 8MB or reduce bandwidth",
+        position = 1003,
+        section = advancedSection
+    )
+    @Units(Units.PERCENT)
+    @Range(min = 1, max = 100)
+    default int screenshotScale() {
+        return 100;
+    }
+
+    @ConfigItem(
         keyName = "discordRichEmbeds",
         name = "Use Rich Embeds",
         description = "Whether Discord's rich embed format should be used for webhooks",
-        position = 1003,
+        position = 1004,
         section = advancedSection
     )
     default boolean discordRichEmbeds() {
@@ -184,7 +198,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "embedFooterText",
         name = "Embed Footer Text",
         description = "The text in the footer of rich embed webhook messages. If empty, no footer will be sent",
-        position = 1004,
+        position = 1005,
         section = advancedSection
     )
     default String embedFooterText() {
@@ -195,7 +209,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "embedFooterIcon",
         name = "Embed Footer Icon",
         description = "The URL for the footer icon image of rich embed webhooks. Requires footer text to not be empty",
-        position = 1005,
+        position = 1006,
         section = advancedSection
     )
     default String embedFooterIcon() {
@@ -206,7 +220,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "ignoredNames",
         name = "Ignored RSNs",
         description = "Prevent notifications from the following player names (One name per line)",
-        position = 1006,
+        position = 1007,
         section = advancedSection
     )
     default String ignoredNames() {
@@ -217,7 +231,7 @@ public interface DinkPluginConfig extends Config {
         keyName = "playerLookupService",
         name = "Player Lookup Service",
         description = "The service used to lookup a players account, to make their name clickable in Discord embeds",
-        position = 1007,
+        position = 1008,
         section = advancedSection
     )
     default PlayerLookupService playerLookupService() {
@@ -482,10 +496,21 @@ public interface DinkPluginConfig extends Config {
     }
 
     @ConfigItem(
+        keyName = "levelNotifyCombat",
+        name = "Notify for Combat Levels",
+        description = "Whether notifications should occur for combat level increases",
+        position = 23,
+        section = levelSection
+    )
+    default boolean levelNotifyCombat() {
+        return false;
+    }
+
+    @ConfigItem(
         keyName = "levelInterval",
         name = "Notify Interval",
         description = "Interval between when a notification should be sent",
-        position = 23,
+        position = 24,
         section = levelSection
     )
     default int levelInterval() {
@@ -497,7 +522,7 @@ public interface DinkPluginConfig extends Config {
         name = "Minimum Skill Level",
         description = "The minimum skill level required to send a notification.<br/>" +
             "Useful for filtering out low-level notifications",
-        position = 24,
+        position = 25,
         section = levelSection
     )
     default int levelMinValue() {
@@ -510,7 +535,7 @@ public interface DinkPluginConfig extends Config {
         description = "The message to be sent through the webhook.<br/>" +
             "Use %USERNAME% to insert your username<br/>" +
             "Use %SKILL% to insert the levelled skill(s)",
-        position = 25,
+        position = 26,
         section = levelSection
     )
     default String levelNotifyMessage() {