+++ src/main/java/dinkplugin/notifiers/LevelNotifier.java
@@ -1,5 +1,6 @@
 package dinkplugin.notifiers;
 
+import com.google.common.collect.ImmutableSet;
 import dinkplugin.message.NotificationBody;
 import dinkplugin.message.NotificationType;
 import dinkplugin.notifiers.data.LevelNotificationData;
@@ -17,6 +18,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.ConcurrentHashMap;
@@ -25,6 +27,8 @@ import java.util.concurrent.atomic.AtomicInteger;
 @Slf4j
 @Singleton
 public class LevelNotifier extends BaseNotifier {
+    private static final String COMBAT_NAME = "Combat";
+    private static final Set<String> COMBAT_COMPONENTS;
     private final BlockingQueue<String> levelledSkills = new ArrayBlockingQueue<>(Skill.values().length);
     private final Map<String, Integer> currentLevels = new ConcurrentHashMap<>();
     private final AtomicInteger ticksWaited = new AtomicInteger();
@@ -42,6 +46,7 @@ public class LevelNotifier extends BaseNotifier {
                 currentLevels.put(skill.getName(), Experience.getLevelForXp(client.getSkillExperience(skill)));
             }
         }
+        currentLevels.put(COMBAT_NAME, calculateCombatLevel());
     }
 
     public void reset() {
@@ -84,18 +89,48 @@ public class LevelNotifier extends BaseNotifier {
         int virtualLevel = level < 99 ? level : Experience.getLevelForXp(xp); // avoid log(n) query when not needed
         Integer previousLevel = currentLevels.put(skill, virtualLevel);
 
-        if (!config.notifyLevel() || previousLevel == null || virtualLevel == previousLevel) {
-            return;
-        }
-
-        if (virtualLevel < previousLevel) {
+        if (previousLevel != null && virtualLevel < previousLevel) {
             // base skill level should never regress; reset notifier state
             reset();
             return;
         }
 
-        if (checkLevelInterval(previousLevel, virtualLevel) && levelledSkills.offer(skill)) {
-            log.debug("Observed level up for {} to {}", skill, virtualLevel);
+        // Check normal skill level up
+        checkLevelUp(config.notifyLevel(), skill, previousLevel, virtualLevel);
+
+        // Skip combat level checking if no level up has occurred
+        if (previousLevel == null || virtualLevel <= previousLevel) {
+            // only return if we don't need to initialize combat level for the first time
+            if (currentLevels.containsKey(COMBAT_NAME))
+                return;
+        }
+
+        // Check for combat level increase
+        if (COMBAT_COMPONENTS.contains(skill)) {
+            int combatLevel = calculateCombatLevel();
+            Integer previousCombatLevel = currentLevels.put(COMBAT_NAME, combatLevel);
+            checkLevelUp(config.levelNotifyCombat(), COMBAT_NAME, previousCombatLevel, combatLevel);
+        }
+    }
+
+    private void checkLevelUp(boolean configEnabled, String skill, Integer previousLevel, int currentLevel) {
+        if (previousLevel == null || currentLevel <= previousLevel) {
+            log.trace("Ignoring non-level-up for {}: {}", skill, currentLevel);
+            return;
+        }
+
+        if (!configEnabled) {
+            log.trace("Ignoring level up of {} to {} due to disabled config setting", skill, currentLevel);
+            return;
+        }
+
+        if (!checkLevelInterval(previousLevel, currentLevel, COMBAT_NAME.equals(skill))) {
+            log.trace("Ignoring level up of {} from {} to {} that does not align with config interval", skill, previousLevel, currentLevel);
+            return;
+        }
+
+        if (levelledSkills.offer(skill)) {
+            log.debug("Observed level up for {} to {}", skill, currentLevel);
 
             // allow more accumulation of level ups into single notification
             ticksWaited.set(0);
@@ -103,12 +138,15 @@ public class LevelNotifier extends BaseNotifier {
     }
 
     private void attemptNotify() {
-        StringBuilder skillMessage = new StringBuilder();
+        // Prepare level state
         List<String> levelled = new ArrayList<>(levelledSkills.size());
         levelledSkills.drainTo(levelled);
         int count = levelled.size();
         Map<String, Integer> lSkills = new HashMap<>(count);
+        Map<String, Integer> currentLevels = new HashMap<>(this.currentLevels);
 
+        // Build skillMessage and populate lSkills
+        StringBuilder skillMessage = new StringBuilder();
         for (int index = 0; index < count; index++) {
             String skill = levelled.get(index);
             if (index > 0) {
@@ -125,26 +163,54 @@ public class LevelNotifier extends BaseNotifier {
             lSkills.put(skill, level);
         }
 
-        String thumbnail = count == 1 ? getSkillIcon(levelled.get(0)) : null;
+        // Separately check for combat level increase for extra data
+        Boolean combatLevelUp = lSkills.remove(COMBAT_NAME) != null; // remove Combat from levelledSkills
+        Integer combatLevel = currentLevels.remove(COMBAT_NAME); // remove Combat from allSkills
+        if (combatLevel == null) {
+            combatLevelUp = null; // combat level was not populated, so it is unclear whether combat level increased
+            combatLevel = calculateCombatLevel(); // populate combat level for extra data
+        } else if (!config.levelNotifyCombat()) {
+            combatLevelUp = null; // if levelNotifyCombat is disabled, it is unclear whether combat level increased
+        }
+        LevelNotificationData.CombatLevel combatData = new LevelNotificationData.CombatLevel(combatLevel, combatLevelUp);
+
+        // Select relevant thumbnail url
+        String thumbnail;
+        if (count == 1) {
+            // Use skill icon if only one skill was levelled up
+            thumbnail = getSkillIcon(levelled.get(0));
+        } else if (combatLevelUp != null && combatLevelUp && count == 2) {
+            // Upon a combat level increase, use icon of the other combat-related skill that was levelled up
+            String skill = levelled.get(0);
+            if (COMBAT_NAME.equals(skill))
+                skill = levelled.get(1);
+            thumbnail = getSkillIcon(skill);
+        } else {
+            // Fall back to NotificationType#getThumbnail
+            thumbnail = null;
+        }
+
+        // Populate message template
         String fullNotification = StringUtils.replaceEach(
             config.levelNotifyMessage(),
             new String[] { "%USERNAME%", "%SKILL%" },
             new String[] { Utils.getPlayerName(client), skillMessage.toString() }
         );
 
+        // Fire notification
         createMessage(config.levelSendImage(), NotificationBody.builder()
             .text(fullNotification)
-            .extra(new LevelNotificationData(lSkills, new HashMap<>(currentLevels)))
+            .extra(new LevelNotificationData(lSkills, currentLevels, combatData))
             .type(NotificationType.LEVEL)
             .thumbnailUrl(thumbnail)
             .build());
     }
 
-    private boolean checkLevelInterval(int previous, int level) {
+    private boolean checkLevelInterval(int previous, int level, boolean skipVirtualCheck) {
         if (level < config.levelMinValue())
             return false;
 
-        if (level > 99 && !config.levelNotifyVirtual())
+        if (!skipVirtualCheck && level > 99 && !config.levelNotifyVirtual())
             return false;
 
         int interval = config.levelInterval();
@@ -157,7 +223,31 @@ public class LevelNotifier extends BaseNotifier {
         return remainder == 0 || (level - remainder) > previous;
     }
 
+    private int calculateCombatLevel() {
+        return Experience.getCombatLevel(
+            client.getRealSkillLevel(Skill.ATTACK),
+            client.getRealSkillLevel(Skill.STRENGTH),
+            client.getRealSkillLevel(Skill.DEFENCE),
+            client.getRealSkillLevel(Skill.HITPOINTS),
+            client.getRealSkillLevel(Skill.MAGIC),
+            client.getRealSkillLevel(Skill.RANGED),
+            client.getRealSkillLevel(Skill.PRAYER)
+        );
+    }
+
     private static String getSkillIcon(String skillName) {
         return Utils.WIKI_IMG_BASE_URL + skillName + "_icon.png";
+    }
+
+    static {
+        COMBAT_COMPONENTS = ImmutableSet.of(
+            Skill.ATTACK.getName(),
+            Skill.STRENGTH.getName(),
+            Skill.DEFENCE.getName(),
+            Skill.HITPOINTS.getName(),
+            Skill.MAGIC.getName(),
+            Skill.RANGED.getName(),
+            Skill.PRAYER.getName()
+        );
     }
 }