+++ src/test/java/dinkplugin/notifiers/MockedNotifierTest.java
@@ -111,6 +111,7 @@ abstract class MockedNotifierTest extends MockedTestBase {
         when(config.maxRetries()).thenReturn(0);
         when(config.baseRetryDelay()).thenReturn(2000);
         when(config.imageWriteTimeout()).thenReturn(30_000);
+        when(config.screenshotScale()).thenReturn(95);
         when(config.discordRichEmbeds()).thenReturn(!"false".equalsIgnoreCase(System.getenv("TEST_WEBHOOK_RICH")));
         when(config.embedFooterText()).thenReturn("Powered by Dink");
         when(config.embedFooterIcon()).thenReturn("https://github.com/pajlads/DinkPlugin/raw/master/icon.png");