+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Minor: Rescale screenshots to comply with Discord's 8MB size limit. (#200)
+- Minor: Include combat level in skill notifications. (#203)
 - Dev: Refactor killer identification in death notifier. (#197)
 
 ## 1.3.2